+++ webextension/js/index.js
@@ -11,8 +11,11 @@ import {sendNotification} from "./classes/chrome-notification.js";
 import {ChromeUpdateNotification} from './classes/chromeUpdateNotification.js';
 
 import './variousFeatures/tabPageServerIp.js';
+import './devtools/devtools-background.js';
 
 import {isFirefox} from "./utils/browserDetect.js";
+import "./newTab/newTab-background.js";
+import {onDevToolsJson} from "./devtools/devtools-background.js";
 if (isFirefox) {
 	import('./variousFeatures/copyTextLink.js')
 		.catch(console.error)
@@ -52,6 +55,9 @@ chrome.runtime.onMessage.addListener(function (message, sender, sendResponse) {
 						});
 					})
 				;
+				return true;
+			case "ztoolbox_devtools_json":
+				onDevToolsJson(message.data);
 				return true;
 		}
 	}