+++ webextension/js/variousFeatures/newTab-init.js
@@ -1,4 +1,0 @@
-import('./newTab.js')
-	.catch(console.error)
-;
-document.querySelector('#newTab-script')?.remove();