+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+# 2.2.0
+* \+ : Capture page system for the new tab page
+* \+ : Restore last closed tab button for the new tab page
+* \+ : JSON view test in devtools (Chrome Only)
+* Fix : Blank with new tab page  
+
 # 2.1.0
 * i: Move from `Material Icons` to `Material Symbols`
+* \+ : New tab page
 * Fix: Better Vivaldi detection for WebSocket (using speed dial and bookmark bar properties)
 
 # 2.0.2